rework inverse index creation
Used to store only the first position of a given word in the document,
but try do so multiple times (which would fail given the primary key
constraint.
Removed the frequency column since it is redundant. Also replaced the
primary key constraint with a UNIQUE over document_id and position.

Unfortunately, this version is too slow; it only managed to index 83
documents in ~87 minutes.