make the crawler multi-threaded
Unfortunately, it did not work to have the threads compete for the same
database file(s), so a lot had to be changed.
Worker threads (processes actually) receive work from the main crawler
process, which coordinates (and serializes) all database accesses. It
takes the next URL off the queue, manages the rate-limiting "tokens",
stores the robots.txt's and requests (which the workers fetch). Workers
calculate the relevance & simhash and extract the links, but the main
process still has to search the entire document list for nearly similar
simhashes.